requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
fake-useragent==1.4.0
python-telegram-bot==20.3